Better build.sh

+++ build.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
 if [ -z $1 ]; then
+	if [ -z $KBUILD_BUILD_VERSION ]; then
+		export KBUILD_BUILD_VERSION="Test-`date '+%Y%m%d-%H%m'`"
+	fi
 	echo "Kernel will be labelled ($KBUILD_BUILD_VERSION)"
 else
 	echo "Setting kernel name to ($1)"
@@ -8,11 +11,16 @@ else
 fi
 
 echo "Compiling the kernel"
+rm arch/arm/boot/zImage
 make -j2
 
-echo "Tarballing the kernel"
-cp arch/arm/boot/zImage ./
-tar cf $KBUILD_BUILD_VERSION-zImage.tar zImage
-rm zImage
+if test -f arch/arm/boot/zImage; then
+	echo "Tarballing the kernel"
+	cp arch/arm/boot/zImage ./
+	tar cf $KBUILD_BUILD_VERSION-zImage.tar zImage
+	rm zImage
+else
+	echo "Will not tarball as make didn't produce zImage"
+fi
 
 echo "Done"
